requirements.txt
azure-cognitiveservices-speech==1.24.0
PyQt6==6.4.0
PyPDF2==2.11.1
